pyproject.toml
[project]
name = "mc_arc"
version = "0.1.0"
description = "Experimental multi-agent conversation framework enabling natural group discussions between AI agents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "anthropic>=0.52.1",
    "google-genai>=1.17.0",
    "lancedb>=0.24.0",
    "logfire>=3.18.0",
    "openai>=1.82.1",
    "pyarrow>=20.0.0",
    "pydantic-ai>=0.2.16",
    "python-dotenv>=1.1.0",
]
